ali_1_final/ali1_ranking/requirements.txt
absl-py==1.0.0
alabaster==0.7.11
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
blessings==1.7
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2022.5.18
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
conda==4.5.11
conda-build==3.15.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dask==0.19.1
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.23.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
faiss-gpu==1.7.2
fastcache==1.0.2
filelock==3.7.0
Flask==1.0.2
Flask-Cors==3.0.6
gast==0.5.3
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
gpustat==0.6.0
greenlet==0.4.15
grpcio==1.45.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
huggingface-hub==0.6.0
hyperlink==18.0.0
idna==3.3
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==4.11.3
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython_genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10
joblib==1.1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.24.0
locket==0.2.0
lxml==4.2.5
Markdown==3.3.6
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
mock==4.0.3
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.21.6
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.6
packaging==21.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkginfo==1.4.2
pluggy==0.7.1
ply==3.11
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.20.1
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==2.0.3
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==3.0.9
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
PyYAML==6.0
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
regex==2022.4.24
requests==2.27.1
rope==0.11.0
ruamel_yaml==0.15.46
sacremoses==0.0.53
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tokenizers==0.12.1
toolz==0.9.0
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tornado==5.1
tqdm==4.64.0
traitlets==4.3.2
transformers==4.17.0
Twisted==18.7.0
typing_extensions==4.2.0
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
zict==0.1.3
zipp==3.8.0
zope.interface==4.5.0
